+++ flamenco/eve_settings.py
@@ -293,6 +293,7 @@ tasks_schema = {
             'blender-render',  # Rendering with Blender.
             'exr-merge',  # EXR merging, probably also happens with Blender.
             'file-management',  # removing directory trees, moving files around, etc.
+            'video-encoding',  # Running things through ffmpeg to produce videos.
         ]
     },
     'commands': {