+++ tests/test_job_compilers.py
@@ -6,6 +6,8 @@ from unittest import mock
 
 from bson import ObjectId, tz_util
 
+from abstract_flamenco_test import AbstractFlamencoTest
+
 
 class JobDocForTesting(dict):
     """Dict that doesn't show the contents in its repr().
@@ -130,10 +132,15 @@ class CommandTest(unittest.TestCase):
         }, cmd.to_dict())
 
 
-class BlenderRenderTest(unittest.TestCase):
+class BlenderRenderTest(AbstractFlamencoTest):
     def setUp(self):
+        super().setUp()
+
         # Create a timestamp before we start mocking datetime.datetime.
         self.created = datetime.datetime(2018, 7, 6, 11, 52, 33, tzinfo=tz_util.utc)
+
+        mngr_doc, account, token = self.create_manager_service_account()
+        self.mngr_id = mngr_doc['_id']
 
     def test_intermediate_path(self):
         from flamenco.job_compilers import blender_render
@@ -240,6 +247,205 @@ class BlenderRenderTest(unittest.TestCase):
         task_manager.api_set_task_status_for_job.assert_called_with(
             job_doc['_id'], 'under-construction', 'queued', now=mock_now)
         job_manager.api_set_job_status(job_doc['_id'], 'under-construction', 'queued', now=mock_now)
+
+    def test_create_video(self):
+        from flamenco.job_compilers import blender_render, commands
+
+        with self.app.app_context():
+            self.flamenco.db('managers').update_one(
+                {'_id': self.mngr_id},
+                {'$set': {'worker_task_types': ['blender-render', 'video-encoding']}}
+            )
+
+        job_doc = JobDocForTesting({
+            '_id': ObjectId(24 * 'f'),
+            '_created': self.created,
+            'manager': self.mngr_id,
+            'settings': {
+                'frames': '1-5',
+                'chunk_size': 3,
+                'render_output': '/render/out/frames-######',
+                'format': 'EXR',
+                'filepath': '/agent327/scenes/someshot/somefile.flamenco.blend',
+                'blender_cmd': '/path/to/blender --enable-new-depsgraph',
+
+                # On top of pretty much the same settings as test_small_job(),
+                # we add those settings that trigger the creation of the
+                # create_video task.
+                'fps': 24,
+                'images_or_video': 'images',
+                'output_file_extension': '.exr',
+            },
+            'job_type': 'blender-render',
+        })
+
+        task_manager = mock.Mock()
+        job_manager = mock.Mock()
+
+        # We expect:
+        # - 2 chunk of 3 resp 2 frames.
+        # - 1 create_video task.
+        # - 1 move-to-final task.
+        # so that's 4 tasks in total.
+        task_ids = [ObjectId() for _ in range(4)]
+        task_manager.api_create_task.side_effect = task_ids
+
+        compiler = blender_render.BlenderRender(
+            task_manager=task_manager, job_manager=job_manager)
+
+        with self.app.app_context():
+            compiler.compile(job_doc)
+
+        task_manager.api_create_task.assert_has_calls([
+            # Render tasks
+            mock.call(
+                job_doc,
+                [commands.BlenderRender(
+                    blender_cmd='/path/to/blender --enable-new-depsgraph',
+                    filepath='/agent327/scenes/someshot/somefile.flamenco.blend',
+                    format='EXR',
+                    render_output='/render/out__intermediate-2018-07-06_115233/frames-######',
+                    frames='1..3')],
+                'blender-render-1-3',
+                status='under-construction',
+                task_type='blender-render',
+            ),
+            mock.call(
+                job_doc,
+                [commands.BlenderRender(
+                    blender_cmd='/path/to/blender --enable-new-depsgraph',
+                    filepath='/agent327/scenes/someshot/somefile.flamenco.blend',
+                    format='EXR',
+                    render_output='/render/out__intermediate-2018-07-06_115233/frames-######',
+                    frames='4,5')],
+                'blender-render-4,5',
+                status='under-construction',
+                task_type='blender-render',
+            ),
+
+            # Create a video of the final frames.
+            mock.call(
+                job_doc,
+                [commands.CreateVideo(
+                    input_files='/render/out__intermediate-2018-07-06_115233/*.exr',
+                    output_file='/render/out__intermediate-2018-07-06_115233/somefile-1-5.mkv',
+                    fps=24,
+                )],
+                'create-video',
+                parents=task_ids[0:2],
+                status='under-construction',
+                task_type='video-encoding',
+            ),
+
+            # Move to final location
+            mock.call(
+                job_doc,
+                [commands.MoveToFinal(
+                    src='/render/out__intermediate-2018-07-06_115233',
+                    dest='/render/out')],
+                'move-to-final',
+                parents=[task_ids[2]],
+                status='under-construction',
+                task_type='file-management',
+            ),
+        ])
+
+        task_manager.api_set_task_status_for_job.assert_called_with(
+            job_doc['_id'], 'under-construction', 'queued', now=mock.ANY)
+        job_manager.api_set_job_status(job_doc['_id'], 'under-construction', 'queued', now=mock.ANY)
+
+    def create_video_without_proper_task_type_support(self):
+        from flamenco.job_compilers import blender_render, commands
+
+        with self.app.app_context():
+            self.flamenco.db('managers').update_one(
+                {'_id': self.mngr_id},
+                # No video-encoding task type
+                {'$set': {'worker_task_types': ['blender-render']}}
+            )
+
+        job_doc = JobDocForTesting({
+            '_id': ObjectId(24 * 'f'),
+            '_created': self.created,
+            'manager': self.mngr_id,
+            'settings': {
+                'frames': '1-5',
+                'chunk_size': 3,
+                'render_output': '/render/out/frames-######',
+                'format': 'EXR',
+                'filepath': '/agent327/scenes/someshot/somefile.flamenco.blend',
+                'blender_cmd': '/path/to/blender --enable-new-depsgraph',
+
+                # On top of pretty much the same settings as test_small_job(),
+                # we add those settings that trigger the creation of the
+                # create_video task.
+                'fps': 24,
+                'images_or_video': 'images',
+                'output_file_extension': '.exr',
+            },
+            'job_type': 'blender-render',
+        })
+
+        task_manager = mock.Mock()
+        job_manager = mock.Mock()
+
+        # We expect:
+        # - 2 chunk of 3 resp 2 frames.
+        # - 1 create_video task.
+        # - 1 move-to-final task.
+        # so that's 4 tasks in total.
+        task_ids = [ObjectId() for _ in range(4)]
+        task_manager.api_create_task.side_effect = task_ids
+
+        compiler = blender_render.BlenderRender(
+            task_manager=task_manager, job_manager=job_manager)
+
+        with self.app.app_context():
+            compiler.compile(job_doc)
+
+        task_manager.api_create_task.assert_has_calls([
+            # Render tasks
+            mock.call(
+                job_doc,
+                [commands.BlenderRender(
+                    blender_cmd='/path/to/blender --enable-new-depsgraph',
+                    filepath='/agent327/scenes/someshot/somefile.flamenco.blend',
+                    format='EXR',
+                    render_output='/render/out__intermediate-2018-07-06_115233/frames-######',
+                    frames='1..3')],
+                'blender-render-1-3',
+                status='under-construction',
+                task_type='blender-render',
+            ),
+            mock.call(
+                job_doc,
+                [commands.BlenderRender(
+                    blender_cmd='/path/to/blender --enable-new-depsgraph',
+                    filepath='/agent327/scenes/someshot/somefile.flamenco.blend',
+                    format='EXR',
+                    render_output='/render/out__intermediate-2018-07-06_115233/frames-######',
+                    frames='4,5')],
+                'blender-render-4,5',
+                status='under-construction',
+                task_type='blender-render',
+            ),
+
+            # Move to final location
+            mock.call(
+                job_doc,
+                [commands.MoveToFinal(
+                    src='/render/out__intermediate-2018-07-06_115233',
+                    dest='/render/out')],
+                'move-to-final',
+                parents=task_ids[1:2],
+                status='under-construction',
+                task_type='file-management',
+            ),
+        ])
+
+        task_manager.api_set_task_status_for_job.assert_called_with(
+            job_doc['_id'], 'under-construction', 'queued', now=mock.ANY)
+        job_manager.api_set_job_status(job_doc['_id'], 'under-construction', 'queued', now=mock.ANY)
 
 
 class BlenderRenderProgressiveTest(unittest.TestCase):